+++ queosk/src/main/java/com/bttf/queosk/controller/RestaurantController.java
@@ -1,8 +1,7 @@
 package com.bttf.queosk.controller;
 
-import com.bttf.queosk.domain.RestaurantSignInForm;
+import com.bttf.queosk.dto.RestaurantSignInForm;
 import com.bttf.queosk.service.RestaurantService;
-import com.bttf.queosk.util.KakaoGeoAddress;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
