+++ queosk/src/main/java/com/bttf/queosk/dto/LoginType.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.domain.enumerate;
+package com.bttf.queosk.dto;
 
 public enum LoginType {
     NORMAL, KAKAO
