+++ queosk/src/main/java/com/bttf/queosk/dto/MenuStatus.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.domain.enumerate;
+package com.bttf.queosk.dto;
 
 public enum MenuStatus {
     SOLD_OUT, ON_SALE
