+++ queosk/src/main/java/com/bttf/queosk/dto/OperationStatus.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.domain.enumerate;
+package com.bttf.queosk.dto;
 
 public enum OperationStatus {
     OPEN,
