+++ queosk/src/main/java/com/bttf/queosk/dto/OrderStatus.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.domain.enumerate;
+package com.bttf.queosk.dto;
 
 public enum OrderStatus {
     IN_PROGRESS, DONE, CANCELED
