+++ queosk/src/main/java/com/bttf/queosk/dto/RestaurantCategory.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.domain.enumerate;
+package com.bttf.queosk.dto;
 
 public enum RestaurantCategory {
     DONKASU_SUSHI_JAPANESE,
