+++ queosk/src/main/java/com/bttf/queosk/dto/UserStatus.java
@@ -1,4 +1,4 @@
-package com.bttf.queosk.domain.enumerate;
+package com.bttf.queosk.dto;
 
 public enum UserStatus {
     VERIFIED,
