+++ queosk/src/test/java/com/bttf/queosk/repository/OrderRepositoryTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
-import static com.bttf.queosk.domain.enumerate.TableStatus.OPEN;
+import static com.bttf.queosk.dto.TableStatus.OPEN;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @Import(JpaAuditingConfiguration.class)
