+++ queosk/src/test/java/com/bttf/queosk/repository/QueueRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.bttf.queosk.repository;
 
-import com.bttf.queosk.domain.enumerate.UserStatus;
+import com.bttf.queosk.dto.UserStatus;
 import com.bttf.queosk.entity.QueueEntity;
 import com.bttf.queosk.entity.RestaurantEntity;
 import com.bttf.queosk.entity.UserEntity;
